[REF] html_editor: split link tests to different files
Before this commit: link related tests are mostly inside popover.test.js
because the popover handles the link's functionality.

After this commit: we split the popover tests to five different files:
1. popover.test.js: tests how the popover show react, what content it
shows.
2. create.test.js: tests the creation of a link.
3. edit.test.js: tests the edit/update of a link.
4. transform.test.js: tests the auto transformation of url inserting.
5. image_link.test.js: tests an image with a url

Note that a redundant test is removed: `Should properly show the preview if fetching metadata fails`

task-5974221
