+++ addons/html_editor/static/tests/link/popover.test.js
@@ -1,15 +1,5 @@
 import { describe, expect, test } from "@odoo/hoot";
-import {
-    click,
-    fill,
-    press,
-    queryAll,
-    queryFirst,
-    queryOne,
-    queryText,
-    waitFor,
-    waitForNone,
-} from "@odoo/hoot-dom";
+import { click, fill, press, queryFirst, queryOne, waitFor, waitForNone } from "@odoo/hoot-dom";
 import { animationFrame, tick } from "@odoo/hoot-mock";
 import { markup } from "@odoo/owl";
 import { contains, dataURItoBlob, onRpc, patchWithCleanup } from "@web/../tests/web_test_helpers";
@@ -17,12 +7,9 @@ import { setupEditor } from "../_helpers/editor";
 import { cleanLinkArtifacts } from "../_helpers/format";
 import { getContent, setContent, setSelection } from "../_helpers/selection";
 import { expectElementCount } from "../_helpers/ui_expectations";
-import { insertLineBreak, insertText, splitBlock, undo } from "../_helpers/user_actions";
+import { insertText } from "../_helpers/user_actions";
 import { execCommand } from "../_helpers/userCommands";
 import { MAIN_PLUGINS, NO_EMBEDDED_COMPONENTS_FALLBACK_PLUGINS } from "@html_editor/plugin_sets";
-
-const base64Img =
-    "data:image/png;base64, iVBORw0KGgoAAAANSUhEUgAAAAUA\n        AAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO\n            9TXL0Y4OHwAAAABJRU5ErkJggg==";
 
 describe("should open a popover", () => {
     test("should open a popover when the selection is inside a link and close outside of a link", async () => {
@@ -354,6 +341,33 @@ describe("popover should edit,copy,remove the link", () => {
             '<p>this is a <a href="http://test.com/">http://test.com/[]</a></p>'
         );
     });
+    test("should trap focus within link popover when using Tab and Shift+Tab", async () => {
+        await setupEditor(`<p><a>li[]nk</a></p>`);
+        await expectElementCount(".o-we-linkpopover", 1);
+
+        await animationFrame();
+        expect(".o_we_href_input_link").toBeFocused();
+        // Tab through all focusable elements
+        await press("Tab");
+        await animationFrame();
+        expect("button[name='link_type']").toBeFocused();
+        await press("Tab");
+        await animationFrame();
+        expect(".form-check-input").toBeFocused();
+        await press("Tab");
+        await animationFrame();
+        expect(".o_we_discard_link").toBeFocused();
+
+        // One more Tab should wrap to first element
+        await press("Tab");
+        await animationFrame();
+        expect(".o_we_label_link").toBeFocused();
+
+        // Shift+Tab should wrap to Last element
+        await press(["Shift", "Tab"]);
+        await animationFrame();
+        expect(".o_we_discard_link").toBeFocused();
+    });
     test("link popover should autoconvert an absolute URL if it's in the domain", async () => {
         onRpc("/html_editor/link_preview_internal", () => ({}));
         onRpc("/contactus", () => ({}));
@@ -387,772 +401,9 @@ describe("popover should edit,copy,remove the link", () => {
             '<p>this is a <a href="/contactus">li[]nk</a></p>'
         );
     });
-    test("Should properly show the preview if fetching metadata fails", async () => {
-        const id = Math.random().toString();
-        onRpc("/html_editor/link_preview_internal", () =>
-            Promise.reject(new Error(`No data ${id}`))
-        );
-        onRpc("/contactus", () => ({}));
-        const originalConsoleWarn = console.warn.bind(console);
-        patchWithCleanup(console, {
-            warn: (msg, error, ...args) => {
-                if (!error?.message?.includes?.(id)) {
-                    originalConsoleWarn(msg, error, ...args);
-                }
-            },
-        });
-        const { el } = await setupEditor('<p><a href="/contactus">a[]b</a></p>');
-        await waitFor(".o-we-linkpopover");
-        expect(cleanLinkArtifacts(getContent(el))).toBe('<p><a href="/contactus">a[]b</a></p>');
-    });
 });
 
-describe("Incorrect URL should be corrected", () => {
-    test("when edit a link's URL to 'test.com', the link's URL should be corrected", async () => {
-        const { el } = await setupEditor('<p>this is a <a href="http://test.com/">li[]nk</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("newtest.com");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>this is a <a href="https://newtest.com">li[]nk</a></p>'
-        );
-    });
-    test("when a link's URL is an email, the link's URL should start with mailto:", async () => {
-        const { el } = await setupEditor("<p>this is a <a>li[]nk</a></p>");
-
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("test@test.com");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>this is a <a href="mailto:test@test.com">li[]nk</a></p>'
-        );
-    });
-    test("when a link's URL is an phonenumber, the link's URL should start with tel:", async () => {
-        const { el } = await setupEditor("<p>this is a <a>li[]nk</a></p>");
-
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("+1234567890");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>this is a <a href="tel:+1234567890">li[]nk</a></p>'
-        );
-    });
-});
-
-describe("Link creation", () => {
-    describe("Creation by space", () => {
-        test("typing valid URL + space should convert to link", async () => {
-            const { editor, el } = await setupEditor("<p>[]</p>");
-            await insertText(editor, "http://google.co.in");
-            await insertText(editor, " ");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://google.co.in">http://google.co.in</a>&nbsp;[]</p>'
-            );
-        });
-        test("typing valid URL without protocol + space should convert to https link", async () => {
-            const { editor, el } = await setupEditor("<p>[]</p>");
-            await insertText(editor, "google.com");
-            await insertText(editor, " ");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="https://google.com">google.com</a>&nbsp;[]</p>'
-            );
-        });
-        test("typing valid http URL + space should convert to http link", async () => {
-            const { editor, el } = await setupEditor("<p>[]</p>");
-            await insertText(editor, "http://google.com");
-            await insertText(editor, " ");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://google.com">http://google.com</a>&nbsp;[]</p>'
-            );
-        });
-        test("typing invalid URL + space should not convert to link", async () => {
-            const { editor, el } = await setupEditor("<p>[]</p>");
-            await insertText(editor, "www.odoo");
-            await insertText(editor, " ");
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>www.odoo []</p>");
-        });
-    });
-    describe("Creation by powerbox", () => {
-        test("click on link command in powerbox should not (yet) create a link element and open the linkpopover", async () => {
-            const { editor, el } = await setupEditor("<p>ab[]</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            expect(".active .o-we-command-name").toHaveText("Link");
-
-            await click(".o-we-command-name:first");
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>ab[]</p>");
-            await expectElementCount(".o-we-linkpopover", 1);
-            expect(".o-we-linkpopover input.o_we_label_link").toBeFocused({
-                message: "should focus label input by default, when we don't have a label",
-            });
-        });
-
-        test("when create a new link by powerbox and not input anything, the link should be removed", async () => {
-            const { editor, el } = await setupEditor("<p>ab[]</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await expectElementCount(".o-we-linkpopover", 1);
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>ab</p>");
-            // simulate click outside
-            const pNode = queryOne("p");
-            setSelection({
-                anchorNode: pNode,
-                anchorOffset: 0,
-                focusNode: pNode,
-                focusOffset: 0,
-            });
-            await animationFrame();
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>[]ab</p>");
-        });
-        test("when create a new link by powerbox and not input anything, the apply link button should be disable", async () => {
-            const { editor, el } = await setupEditor("<p>ab[]</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await waitFor(".o-we-linkpopover");
-            await expectElementCount(".o-we-linkpopover", 1);
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>ab</p>");
-            expect(".o_we_apply_link").toHaveAttribute("disabled");
-        });
-        test("when create a new link by powerbox and only input the URL, the link should be created with corrected https URL", async () => {
-            const { editor, el } = await setupEditor("<p>ab[]</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill("test.com");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>ab<a href="https://test.com">test.com[]</a></p>'
-            );
-        });
-        test("Should be able to insert link on empty p", async () => {
-            const { editor, el } = await setupEditor("<p>[]<br></p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://test.test/">http://test.test/[]</a></p>'
-            );
-        });
-        test("Should be able to insert button on empty p", async () => {
-            const { editor, el } = await setupEditor("<p>[]<br></p>");
-            await insertText(editor, "/button");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://test.test/" class="btn btn-primary">http://test.test/[]</a></p>'
-            );
-        });
-        test("Should keep http protocol on valid http url", async () => {
-            const { editor, el } = await setupEditor("<p>[]<br></p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://google.com"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://google.com">http://google.com[]</a></p>'
-            );
-        });
-        test("should insert a link and preserve spacing", async () => {
-            const { editor, el } = await setupEditor("<p>a []&nbsp;&nbsp;b</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await contains(".o-we-linkpopover input.o_we_label_link").fill("link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a <a href="http://test.test/">link[]</a>&nbsp;&nbsp;b</p>'
-            );
-        });
-        test("should insert a link then create a new <p>", async () => {
-            const { editor, el } = await setupEditor("<p>ab[]</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await contains(".o-we-linkpopover input.o_we_label_link").fill("link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            // press("Enter");
-            splitBlock(editor);
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                `<p>ab<a href="http://test.test/">link</a></p><p o-we-hint-text='Type "/" for commands' class="o-we-hint">[]<br></p>`
-            );
-        });
-        test("should insert a link then create a new <p>, and another character", async () => {
-            const { editor, el } = await setupEditor("<p>a[]b</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await contains(".o-we-linkpopover input.o_we_label_link").fill("link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            // press("Enter");
-            splitBlock(editor);
-            await insertText(editor, "D");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a<a href="http://test.test/">link</a></p><p>D[]b</p>'
-            );
-        });
-        test("should insert a link then insert a <br>", async () => {
-            const { editor, el } = await setupEditor("<p>a[]b</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await contains(".o-we-linkpopover input.o_we_label_link").fill("link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            // press("Enter");
-            insertLineBreak(editor);
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a<a href="http://test.test/">link</a><br>[]b</p>'
-            );
-        });
-        test("should insert a link then insert a <br> and another character", async () => {
-            const { editor, el } = await setupEditor("<p>a[]b</p>");
-            await insertText(editor, "/link");
-            await animationFrame();
-            await click(".o-we-command-name:first");
-            await contains(".o-we-linkpopover input.o_we_label_link").fill("link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").fill(
-                "http://test.test/"
-            );
-            // press("Enter");
-            insertLineBreak(editor);
-            await insertText(editor, "D");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a<a href="http://test.test/">link</a><br>D[]b</p>'
-            );
-        });
-    });
-    describe("Creation by toolbar", () => {
-        test("should convert all selected text to link", async () => {
-            const { el } = await setupEditor("<p>[Hello]</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://test.test/">Hello[]</a></p>'
-            );
-        });
-        test("discard should close the popover (in iframe)", async () => {
-            await setupEditor("<p>[Hello]</p>", { props: { iframe: true } });
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            await click(".o_we_discard_link");
-            await animationFrame();
-            expect(".o-we-linkpopover").toHaveCount(0);
-        });
-        test("should convert valid url to https link", async () => {
-            const { el } = await setupEditor("<p>[Hello]</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "google.com"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="https://google.com">Hello[]</a></p>'
-            );
-        });
-        test("should convert valid http url to http link", async () => {
-            const { el } = await setupEditor("<p>[Hello]</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "http://google.com"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://google.com">Hello[]</a></p>'
-            );
-        });
-        test("should convert all selected text to link and keep style except color", async () => {
-            const { el } = await setupEditor(
-                '<p>Hello this is [a <b>new</b> <u>link</u> <span style="color:red">keeping</span> style]!</p>'
-            );
-            await waitFor(".o-we-toolbar");
-            click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>Hello this is <a href="http://test.test/">a <b>new</b> <u>link</u> keeping style[]</a>!</p>'
-            );
-        });
-        test("should convert all selected text to link and keep style except bgclor", async () => {
-            const { el } = await setupEditor(
-                '<p>Hello this is [a <b>new</b> <u>link</u> <span style="background-color:red">keeping</span> style]!</p>'
-            );
-            await waitFor(".o-we-toolbar");
-            click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>Hello this is <a href="http://test.test/">a <b>new</b> <u>link</u> keeping style[]</a>!</p>'
-            );
-        });
-        test("should convert all selected text to link and keep style except color (2)", async () => {
-            const { el } = await setupEditor(
-                '<p>Hello this is a <b>ne[w</b> <u>link</u> <span style="color:red">keep]ing</span> style!</p>'
-            );
-            await waitFor(".o-we-toolbar");
-            click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>Hello this is a <b>ne</b><a href="http://test.test/"><b>w</b> <u>link</u> keep[]</a><span style="color:red">ing</span> style!</p>'
-            );
-        });
-        test("should set the link on two existing characters", async () => {
-            const { el } = await setupEditor("<p>H[el]lo</p>");
-            await waitFor(".o-we-toolbar");
-            // link button should be enabled
-            expect('.o-we-toolbar button[name="link"]').not.toHaveClass("disabled");
-            expect('.o-we-toolbar button[name="link"]').not.toHaveAttribute("disabled");
-            await click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>H<a href="http://test.test/">el[]</a>lo</p>'
-            );
-        });
-        test("should not allow to create a link if selection span multiple block", async () => {
-            const { el } = await setupEditor("<p>H[ello</p><p>wor]ld</p>");
-            await waitFor(".o-we-toolbar");
-            // link button should be disabled
-            expect('.o-we-toolbar button[name="link"]').toHaveClass("disabled");
-            expect('.o-we-toolbar button[name="link"]').toHaveAttribute("disabled");
-            await click('.o-we-toolbar button[name="link"]');
-            expect(getContent(el)).toBe("<p>H[ello</p><p>wor]ld</p>");
-        });
-        test("when you open link popover, url input should be focus by default", async () => {
-            const { el } = await setupEditor("<p>[Hello]</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            expect(".o-we-linkpopover input.o_we_href_input_link").toBeFocused();
-
-            await fill("test.com");
-            await waitFor(".o_we_apply_link:not([disabled])");
-            await click(".o_we_apply_link");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="https://test.com">Hello[]</a></p>'
-            );
-        });
-        test("when you open image link popover, url input should be focus by default", async () => {
-            const { el } = await setupEditor(`<p>[<img src="${base64Img}">]</p>`);
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            expect(".o-we-linkpopover input.o_we_href_input_link").toBeFocused();
-
-            await press("escape");
-            await expectElementCount(".o-we-linkpopover", 0);
-            expect(getContent(el)).toBe(`<p>[<img src="${base64Img}">]</p>`);
-        });
-        test("should be correctly unlink/link", async () => {
-            const { el } = await setupEditor('<p>aaaa[b<a href="http://test.com/">cd</a>e]f</p>');
-            await waitFor(".o-we-toolbar");
-
-            await click(".o-we-toolbar .fa-unlink");
-            await tick();
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>aaaa[bcde]f</p>");
-            const pNode = queryOne("p");
-            setSelection({
-                anchorNode: pNode.childNodes[1],
-                anchorOffset: 1,
-                focusNode: pNode.childNodes[0],
-                focusOffset: 3,
-            });
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>aaa<a href="http://test.test/">ab[]</a>cdef</p>'
-            );
-        });
-        test("should remove link when click away without inputting url", async () => {
-            const { el } = await setupEditor("<p>H[el]lo</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            const pNode = queryOne("p");
-            setSelection({
-                anchorNode: pNode,
-                anchorOffset: 0,
-                focusNode: pNode,
-                focusOffset: 0,
-            });
-            await tick();
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>[]Hello</p>");
-        });
-        test("when selection includes partially a link and click the link icon in toolbar, the link should be extended", async () => {
-            const { el } = await setupEditor('<p>a[b<a href="http://test.com/">c]d</a>ef</p>');
-            await waitFor(".o-we-toolbar");
-
-            await click(".o-we-toolbar .fa-link");
-            await expectElementCount(".o-we-linkpopover", 1);
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a<a href="http://test.com/">bcd[]</a>ef</p>'
-            );
-        });
-        test("when selection includes a link and click the link icon in toolbar, the link should be extended", async () => {
-            const { el } = await setupEditor('<p>a[b<a href="http://test.com/">cd</a>e]f</p>');
-            await waitFor(".o-we-toolbar");
-
-            await click(".o-we-toolbar .fa-link");
-            await expectElementCount(".o-we-linkpopover", 1);
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p>a<a href="http://test.com/">bcde[]</a>f</p>'
-            );
-        });
-        test("when create a link on selection which doesn't include a link, it should create a new one", async () => {
-            await setupEditor(
-                '<p><strong>abc<a href="http://test.com/">de</a>te[st</strong> m]e</p>'
-            );
-            await waitFor(".o-we-toolbar");
-
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            expect(".o_we_label_link").toHaveValue("st m");
-            expect(".o_we_href_input_link").toHaveValue("");
-        });
-        test("create a link and undo it (1)", async () => {
-            const { el, editor } = await setupEditor("<p>[Hello]</p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            // not validated link shouldn't affect the DOM yet
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>[Hello]</p>");
-            await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><a href="http://test.test/">Hello[]</a></p>'
-            );
-
-            undo(editor);
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p>[Hello]</p>");
-        });
-        test("create a link and undo it (2)", async () => {
-            const { el, editor } = await setupEditor("<p><b>[Hello]</b></p>");
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            // not validated link shouldn't affect the DOM yet
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p><b>[Hello]</b></p>");
-            await contains(".o-we-linkpopover input.o_we_href_input_link").edit(
-                "http://test.test/"
-            );
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                '<p><b><a href="http://test.test/">Hello[]</a></b></p>'
-            );
-            undo(editor);
-            expect(cleanLinkArtifacts(getContent(el))).toBe("<p><b>[Hello]</b></p>");
-        });
-        test("extend a link on selection and undo it", async () => {
-            const { el, editor } = await setupEditor(
-                `<p>[<a href="https://www.test.com">Hello</a> my friend]</p>`
-            );
-            await waitFor(".o-we-toolbar");
-            await click(".o-we-toolbar .fa-link");
-            await waitFor(".o-we-linkpopover", { timeout: 1500 });
-            expect(queryFirst(".o-we-linkpopover a").href).toBe("https://www.test.com/");
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                `<p><a href="https://www.test.com">Hello my friend[]</a></p>`
-            );
-
-            undo(editor);
-            expect(cleanLinkArtifacts(getContent(el))).toBe(
-                `<p>[<a href="https://www.test.com">Hello</a> my friend]</p>`
-            );
-        });
-    });
-});
-
-describe.tags("desktop");
-describe("Link formatting in the popover", () => {
-    test("click on link, the link popover should load the current format correctly", async () => {
-        await setupEditor(
-            '<p><a href="http://test.com/" class="btn btn-fill-primary">link2[]</a></p>'
-        );
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-        expect(".o_we_label_link").toHaveValue("link2");
-        expect(".o_we_href_input_link").toHaveValue("http://test.com/");
-        expect(queryOne("button[name='link_type']").textContent).toBe("Button Primary");
-    });
-    test("after changing the link format, the link should be updated", async () => {
-        const { el } = await setupEditor(
-            '<p><a href="http://test.com/" class="btn btn-fill-secondary">link2[]</a></p>'
-        );
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-
-        expect(queryOne("button[name='link_type']").textContent).toBe("Button Secondary");
-
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Primary')");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p><a href="http://test.com/" class="btn btn-fill-primary">link2</a></p>`
-        );
-    });
-    test("after changing the link format, the link should be updated (2)", async () => {
-        const { el } = await setupEditor(
-            '<p><a href="http://test.com/" class="btn btn-secondary">link2[]</a></p>'
-        );
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-
-        expect(queryOne("button[name='link_type']").textContent).toBe("Button Secondary");
-
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Primary')");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="http://test.com/" class="btn btn-primary">link2</a></p>'
-        );
-    });
-    test("after changing the link format, the link should be updated (3)", async () => {
-        const { el } = await setupEditor(
-            '<p><a href="http://test.com/" class="random-css-class btn btn-outline-secondary text-muted">link2[]</a></p>'
-        );
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-
-        expect(queryOne("button[name='link_type']").textContent).toBe("Button Secondary");
-
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Primary')");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p><a href="http://test.com/" class="random-css-class text-muted btn btn-outline-primary">link2</a></p>`
-        );
-    });
-    test("after applying the link format, the link's format should be updated", async () => {
-        const { el } = await setupEditor('<p><a href="http://test.com/">link2[]</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-        expect(queryOne("button[name='link_type']").textContent).toBe("Link");
-
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Secondary')");
-        await animationFrame();
-
-        await click(".o_we_apply_link");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="http://test.com/" class="btn btn-secondary">link2[]</a></p>'
-        );
-    });
-    test("clicking the discard button should revert the link format", async () => {
-        const { el } = await setupEditor('<p><a href="http://test.com/">link1[]</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await contains(".o-we-linkpopover input.o_we_label_link").edit("link2");
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Secondary')");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="http://test.com/" class="btn btn-secondary">link2</a></p>'
-        );
-        await click(".o_we_discard_link");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="http://test.com/">link1[]</a></p>'
-        );
-    });
-    test("should close link popover on discard without input", async () => {
-        const { el, editor } = await setupEditor("<p>[]<br></p>");
-        await insertText(editor, "/link");
-        await animationFrame();
-        await click(".o-we-command-name:first");
-        await animationFrame();
-        await waitFor(".o-we-linkpopover");
-
-        await contains(".o_we_discard_link").click();
-        await expectElementCount(".o-we-linkpopover", 0);
-        expect(getContent(el)).toBe(
-            `<p o-we-hint-text='Type "/" for commands' class="o-we-hint">[]<br></p>`
-        );
-    });
-    test("clicking the discard button should revert the link creation", async () => {
-        const { el } = await setupEditor("<p>[link1]</p>");
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-link");
-
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("#", {
-            confirm: false,
-        });
-
-        await click("button[name='link_type']");
-        await animationFrame();
-        await click(".o-we-link-type-dropdown .dropdown-item:contains('Button Secondary')");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="#" class="btn btn-secondary">link1</a></p>'
-        );
-        await click(".o_we_discard_link");
-        expect(cleanLinkArtifacts(getContent(el))).toBe("<p>[link1]</p>");
-        await animationFrame();
-        await waitForNone(".o-we-linkpopover"); // Popover should be closed.
-        await animationFrame();
-        await waitFor(".o-we-toolbar"); // Toolbar should re open.
-    });
-    test("when no label input, the link should have the content of the url", async () => {
-        const { el, editor } = await setupEditor("<p>ab[]</p>");
-        await insertText(editor, "/link");
-        await animationFrame();
-        await click(".o-we-command-name:first");
-        await waitFor(".o-we-linkpopover");
-
-        queryOne(".o_we_href_input_link").focus();
-        for (const char of "newtest.com") {
-            await press(char);
-        }
-        await animationFrame();
-        await click(".o_we_apply_link");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>ab<a href="https://newtest.com">newtest.com[]</a></p>'
-        );
-    });
-    test("link_type dropdown should show a preview of preset with primary and secondary color.", async () => {
-        const { el } = await setupEditor(`<p><a href="http://test.com/">link1[]</a></p>`);
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await animationFrame();
-        await click("button[name='link_type']");
-        await animationFrame();
-        const dropdownItems = queryAll(".o-we-link-type-dropdown .dropdown-item span");
-        expect(dropdownItems[0]).toHaveStyle({ color: "rgb(0, 143, 140)" });
-        expect(dropdownItems[1]).toHaveClass("btn btn-primary btn-sm");
-        expect(dropdownItems[2]).toHaveClass("btn btn-secondary btn-sm");
-        await click(dropdownItems[1]);
-        await animationFrame();
-        expect("button[name='link_type'] span").toHaveClass("btn btn-primary btn-sm");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p><a href="http://test.com/" class="btn btn-primary">link1</a></p>`
-        );
-    });
-});
-
-describe("shortcut", () => {
-    test("create link shortcut should be at the first", async () => {
-        const { editor } = await setupEditor(`<p>[]</p>`);
-        editor.services.command.add("A test command", () => {}, {
-            hotkey: "alt+k",
-            category: "app",
-        });
-
-        await press(["ctrl", "k"]);
-        await animationFrame();
-        expect(queryText(".o_command_name:first")).toBe("Create link");
-    });
-    test.tags("desktop");
-    test("create a link with shortcut", async () => {
-        const { el } = await setupEditor(`<p>[]<br></p>`);
-        // open odoo command bar
-        await press(["ctrl", "k"]);
-        await waitFor('.o_command span[title="Create link"]');
-        // open link tool
-        await click(".o_command_name:first");
-        await waitFor(".o-we-linkpopover");
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("test.com");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="https://test.com">test.com[]</a></p>'
-        );
-    });
-    test.tags("desktop");
-    test("should be able to create link with ctrl+k and ctrl+k", async () => {
-        const { el } = await setupEditor(`<p>[]<br></p>`);
-        // Open odoo global command bar
-        await press(["ctrl", "k"]);
-        await waitFor('.o_command span[title="Create link"]');
-        // Choose the "create link" command to create a link in the editor
-        await press(["ctrl", "k"]);
-        await waitFor(".o-we-linkpopover");
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("test.com");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="https://test.com">test.com[]</a></p>'
-        );
-    });
-    test("should be able to create link with ctrl+k , and should make link on two existing characters", async () => {
-        const { el } = await setupEditor(`<p>H[el]lo</p>`);
-
-        await press(["ctrl", "k"]);
-        await animationFrame();
-        await press(["ctrl", "k"]);
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("test.com");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>H<a href="https://test.com">el[]</a>lo</p>'
-        );
-    });
-    test("Press enter to apply when create a link", async () => {
-        const { el } = await setupEditor(`<p><a>li[]nk</a></p>`);
-
-        await contains(".o-we-linkpopover input.o_we_href_input_link").fill("test.com");
-        await press("Enter");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p><a href="https://test.com">li[]nk</a></p>'
-        );
-    });
-    test("should trap focus within link popover when using Tab and Shift+Tab", async () => {
-        await setupEditor(`<p><a>li[]nk</a></p>`);
-        await expectElementCount(".o-we-linkpopover", 1);
-
-        await animationFrame();
-        expect(".o_we_href_input_link").toBeFocused();
-        // Tab through all focusable elements
-        await press("Tab");
-        await animationFrame();
-        expect("button[name='link_type']").toBeFocused();
-        await press("Tab");
-        await animationFrame();
-        expect(".form-check-input").toBeFocused();
-        await press("Tab");
-        await animationFrame();
-        expect(".o_we_discard_link").toBeFocused();
-
-        // One more Tab should wrap to first element
-        await press("Tab");
-        await animationFrame();
-        expect(".o_we_label_link").toBeFocused();
-
-        // Shift+Tab should wrap to Last element
-        await press(["Shift", "Tab"]);
-        await animationFrame();
-        expect(".o_we_discard_link").toBeFocused();
-    });
-});
-
-describe("link preview", () => {
+describe("popover should show link preview", () => {
     test("test internal link preview", async () => {
         onRpc("/html_editor/link_preview_internal", () => ({
             description: markup("Test description"),
@@ -1320,6 +571,24 @@ describe("link preview", () => {
         await waitFor(".o-we-linkpopover");
         expect.verifySteps([]);
     });
+    test("Should properly show the preview if fetching metadata fails", async () => {
+        const id = Math.random().toString();
+        onRpc("/html_editor/link_preview_internal", () =>
+            Promise.reject(new Error(`No data ${id}`))
+        );
+        onRpc("/contactus", () => ({}));
+        const originalConsoleWarn = console.warn.bind(console);
+        patchWithCleanup(console, {
+            warn: (msg, error, ...args) => {
+                if (!error?.message?.includes?.(id)) {
+                    originalConsoleWarn(msg, error, ...args);
+                }
+            },
+        });
+        const { el } = await setupEditor('<p><a href="/contactus">a[]b</a></p>');
+        await waitFor(".o-we-linkpopover");
+        expect(cleanLinkArtifacts(getContent(el))).toBe('<p><a href="/contactus">a[]b</a></p>');
+    });
     test("test external metadata cached correctly", async () => {
         const title = "Open Source ERP and CRM | Odoo";
         const description =
@@ -1409,178 +678,23 @@ describe("link preview", () => {
         await waitFor(".o-we-linkpopover");
         expect(".o-we-linkpopover .o_we_url_link").toHaveAttribute("target", "_blank");
     });
-});
-
-describe("link in templates", () => {
-    test("Should not remove a link with t-attf-href", async () => {
-        const { el } = await setupEditor('<p>test<a t-attf-href="/test/1">li[]nk</a></p>');
-
-        await expectElementCount(".o-we-linkpopover", 1);
-        const pNode = queryOne("p");
-        setSelection({
-            anchorNode: pNode,
-            anchorOffset: 0,
-            focusNode: pNode,
-            focusOffset: 0,
-        });
-        await waitForNone(".o-we-linkpopover", { timeout: 1500 });
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>[]test<a t-attf-href="/test/1">link</a></p>'
+    test("Should not show link preview when selection in input/textarea", async () => {
+        await setupEditor(
+            `<p><a href="http://test.com/">a[]b</a><span contenteditable="false" data-oe-protected="true"><input></span></p>`
         );
-    });
-    test("Should not remove a link with t-att-href", async () => {
-        const { el } = await setupEditor('<p>test<a t-att-href="/test/1">li[]nk</a></p>');
-
-        await expectElementCount(".o-we-linkpopover", 1);
-        const pNode = queryOne("p");
-        setSelection({
-            anchorNode: pNode,
-            anchorOffset: 0,
-            focusNode: pNode,
-            focusOffset: 0,
-        });
-        await waitForNone(".o-we-linkpopover", { timeout: 1500 });
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            '<p>[]test<a t-att-href="/test/1">link</a></p>'
-        );
-    });
-});
-
-describe("links with inline image", () => {
-    test("can add link to inline image + text", async () => {
-        const { el } = await setupEditor(`<p>ab[cd<img src="${base64Img}">ef]g</p>`);
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-link");
-        await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-            "http://test.test/"
-        );
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">cd<img src="${base64Img}">ef[]</a>g</p>`
-        );
-    });
-    test("can undo add link to inline image + text", async () => {
-        const { editor, el } = await setupEditor(`<p>ab[cd<img src="${base64Img}">ef]g</p>`);
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-link");
-        await contains(".o-we-linkpopover input.o_we_href_input_link", { timeout: 1500 }).edit(
-            "http://test.test/"
-        );
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">cd<img src="${base64Img}">ef[]</a>g</p>`
-        );
-        undo(editor);
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(`<p>ab[cd<img src="${base64Img}">ef]g</p>`);
-    });
-    test("can remove link from an inline image", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">cd<img src="${base64Img}">ef</a>g</p>`
-        );
-        await click("img");
-        await waitFor(".o-we-toolbar");
-        expect("button[name='unlink']").toHaveCount(1);
-        await click("button[name='unlink']");
-        await animationFrame();
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">cd[</a><img src="${base64Img}"><a href="http://test.test/">]ef</a>g</p>`
-        );
+        await waitFor(".o-we-linkpopover");
+        queryOne("input").focus();
         await expectElementCount(".o-we-linkpopover", 0);
     });
-    test("can remove link from a selection of an inline image + text", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">c[d<img src="${base64Img}">e]f</a>g</p>`
-        );
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-unlink");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">c</a>[d<img src="${base64Img}">e]<a href="http://test.test/">f</a>g</p>`
-        );
-    });
-    test("can remove link from a selection (ltr) with multiple inline images", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">c[d<img src="${base64Img}">e<img src="${base64Img}">f]g</a>h</p>`
-        );
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-unlink");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">c</a>[d<img src="${base64Img}">e<img src="${base64Img}">f]<a href="http://test.test/">g</a>h</p>`
-        );
-    });
-    test("can remove link from a selection (rtl) with multiple inline images", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">c]d<img src="${base64Img}">e<img src="${base64Img}">f[g</a>h</p>`
-        );
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-unlink");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">c</a>]d<img src="${base64Img}">e<img src="${base64Img}">f[<a href="http://test.test/">g</a>h</p>`
-        );
-    });
-    test("can remove link from a selection (ltr) with multiple inline images acrossing different links", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">c[d<img src="${base64Img}">e</a>xx<a href="http://test.test/">f<img src="${base64Img}">g]h</a>i</p>`
-        );
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-unlink");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="http://test.test/">c</a>[d<img src="${base64Img}">exxf<img src="${base64Img}">g]<a href="http://test.test/">h</a>i</p>`
-        );
-    });
-    test("can remove link from a selection (rtl) with multiple inline images acrossing different links", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="#">c]d<img src="${base64Img}">e</a>xx<a href="#">f<img src="${base64Img}">g[h</a>i</p>`
-        );
-        await waitFor(".o-we-toolbar");
-        await click(".o-we-toolbar .fa-unlink");
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>ab<a href="#">c</a>]d<img src="${base64Img}">exxf<img src="${base64Img}">g[<a href="#">h</a>i</p>`
-        );
-    });
-    test("link element should be removed and popover should close when image is deleted from a image link", async () => {
-        const { editor, el } = await setupEditor(
-            `<p>ab<a href="http://test.test/"><img src="${base64Img}"></a>c[]</p>`
-        );
-        await click("img");
-        await waitFor(".o-we-toolbar");
-        await waitFor(".o-we-linkpopover");
-
-        execCommand(editor, "deleteImage");
-        await waitForNone(".o-we-linkpopover", { timeout: 1500 });
-
-        expect(cleanLinkArtifacts(getContent(el))).toBe(`<p>ab[]c</p>`);
-    });
-    test("selecting text and a image with link should not extend the link element", async () => {
-        const { el } = await setupEditor(
-            `<p>ab<a href="http://test.test/">cd<img src="${base64Img}">ef</a>g</p>`
-        );
-        setContent(el, `<p>ab<a href="http://test.test/">c]d<img src="${base64Img}">e[f</a>g</p>`);
-        await waitFor(".o-we-linkpopover", { timeout: 1500 });
-        await waitFor(".o-we-toolbar");
-        setContent(el, `<p>a]b<a href="http://test.test/">cd<img src="${base64Img}">e[f</a>g</p>`);
-        await waitForNone(".o-we-linkpopover", { timeout: 1500 });
-        expect(cleanLinkArtifacts(getContent(el))).toBe(
-            `<p>a]b<a href="http://test.test/">cd<img src="${base64Img}">e[f</a>g</p>`
-        );
-    });
 });
 
-describe("readonly mode", () => {
+describe("popover in contenteditable=false or readonly mode", () => {
     test("popover should not display edit buttons in readonly mode", async () => {
         await setupEditor('<p><a class="o_link_readonly" href="http://test.test/">link[]</a></p>');
         await waitFor(".o-we-linkpopover");
         // Edit button should not be available.
         expect(".o-we-linkpopover .o_we_edit_link").toHaveCount(0);
     });
-    // TODO: need to check with AGE
-    test.todo("popover should not open for not editable image", async () => {
-        await setupEditor(`<a href="#"><img src="${base64Img}" contenteditable="false"></a>`);
-        await click("img");
-        await animationFrame();
-        await expectElementCount(".o-we-linkpopover", 0);
-    });
-});
-
-describe("link in contenteditable=false", () => {
     test("popover should not display remove button if link is in a contenteditable=false", async () => {
         await setupEditor(
             '<p contenteditable="false"><a contenteditable="true" href="http://test.test/">link[]</a></p>'
@@ -1603,7 +717,7 @@ describe("link in contenteditable=false", () => {
     });
 });
 
-describe("upload file via link popover", () => {
+describe("popover for file uploads", () => {
     test("should display upload button when url input is empty", async () => {
         const { editor } = await setupEditor("<p>[]<br></p>", {
             config: { Plugins: [...MAIN_PLUGINS, ...NO_EMBEDDED_COMPONENTS_FALLBACK_PLUGINS] },
@@ -1706,65 +820,7 @@ describe("upload file via link popover", () => {
     });
 });
 
-describe("apply button should be disabled when the URL is empty", () => {
-    test("when URL on link is empty, the apply link button should be disabled (1)", async () => {
-        await setupEditor("<p>this is a <a>li[]nk</a></p>");
-        await waitFor(".o-we-linkpopover");
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
-        await tick();
-        expect(".o_we_apply_link").toHaveAttribute("disabled");
-    });
-    test("when URL on link is empty, the apply link button should be disabled (2)", async () => {
-        await setupEditor('<p>this is a <a href="">li[]nk</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
-        await tick();
-        expect(".o_we_apply_link").toHaveAttribute("disabled");
-    });
-    test("when URL on link is empty, the apply link button should be disabled (3)", async () => {
-        await setupEditor('<p>this is a <a href="http://test.test/">li[]nk</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
-        await tick();
-        expect(".o_we_apply_link").toHaveAttribute("disabled");
-    });
-    test("when edit a link's label and URL to '', the apply link button should be disabled", async () => {
-        await setupEditor('<p>this is a <a href="http://test.com/">li[]nk</a></p>');
-        await waitFor(".o-we-linkpopover");
-        await click(".o_we_edit_link");
-        await contains(".o-we-linkpopover input.o_we_label_link").clear();
-        await contains(".o-we-linkpopover input.o_we_href_input_link").clear();
-        await tick();
-        expect(".o_we_apply_link").toHaveAttribute("disabled");
-    });
-});
-
-describe("hidden label field", () => {
-    test("label field should be hidden if <a> content is not text only", async () => {
-        await setupEditor(`<p><a href="http://test.com/"><img src="${base64Img}">te[]xt</a></p>`);
-        await expectElementCount(".o-we-linkpopover", 1);
-        // open edit mode and check if label input is hidden
-        await click(".o_we_edit_link");
-        await waitFor(".o_we_href_input_link", { timeout: 1500 });
-        expect(".o_we_label_link").not.toBeVisible();
-        expect(".o_we_href_input_link").toHaveValue("http://test.com/");
-    });
-});
-
-describe("Selection in input/textarea", () => {
-    test("Should not show link preview", async () => {
-        await setupEditor(
-            `<p><a href="http://test.com/">a[]b</a><span contenteditable="false" data-oe-protected="true"><input></span></p>`
-        );
-        await waitFor(".o-we-linkpopover");
-        queryOne("input").focus();
-        await expectElementCount(".o-we-linkpopover", 0);
-    });
-});
-
-describe("link popover with empty URL", () => {
+describe("popover with empty URL", () => {
     test("should not close the popover when pressing Enter with an empty URL", async () => {
         const { editor } = await setupEditor("<p>ab[]</p>");
         await insertText(editor, "/link");
@@ -1852,32 +908,41 @@ describe("link popover with empty URL", () => {
         await waitForNone(".o-we-linkpopover", { timeout: 1500 });
         expect(cleanLinkArtifacts(getContent(el))).toBe('<p>[]this is a <a href="#">link</a></p>');
     });
-    test("when clearing link URL for an image and clicking outside does not change href", async () => {
-        const { el } = await setupEditor(
-            `<p><a href="http://test.test/">[<img src="${base64Img}">]</a></p>`
-        );
-        await waitFor(".o-we-linkpopover", { timeout: 1500 });
+    test("when URL on link is empty, the apply link button should be disabled (1)", async () => {
+        await setupEditor("<p>this is a <a>li[]nk</a></p>");
+        await waitFor(".o-we-linkpopover");
+        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
+        await tick();
+        expect(".o_we_apply_link").toHaveAttribute("disabled");
+    });
+    test("when URL on link is empty, the apply link button should be disabled (2)", async () => {
+        await setupEditor('<p>this is a <a href="">li[]nk</a></p>');
+        await waitFor(".o-we-linkpopover");
         await click(".o_we_edit_link");
-        await waitFor(".o_we_href_input_link");
+        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
+        await tick();
+        expect(".o_we_apply_link").toHaveAttribute("disabled");
+    });
+    test("when URL on link is empty, the apply link button should be disabled (3)", async () => {
+        await setupEditor('<p>this is a <a href="http://test.test/">li[]nk</a></p>');
+        await waitFor(".o-we-linkpopover");
+        await click(".o_we_edit_link");
+        await contains(".o-we-linkpopover input.o_we_href_input_link").edit("");
+        await tick();
+        expect(".o_we_apply_link").toHaveAttribute("disabled");
+    });
+    test("when edit a link's label and URL to '', the apply link button should be disabled", async () => {
+        await setupEditor('<p>this is a <a href="http://test.com/">li[]nk</a></p>');
+        await waitFor(".o-we-linkpopover");
+        await click(".o_we_edit_link");
+        await contains(".o-we-linkpopover input.o_we_label_link").clear();
         await contains(".o-we-linkpopover input.o_we_href_input_link").clear();
-        await click(el);
-        // Simulate click outside
-        const pNode = queryOne("p");
-        setSelection({
-            anchorNode: pNode,
-            anchorOffset: 0,
-            focusNode: pNode,
-            focusOffset: 0,
-        });
-        await tick(); // wait for selection change.
-        await waitForNone(".o-we-linkpopover", { timeout: 1500 });
-        expect(getContent(el)).toBe(
-            `<p>[]<a href="http://test.test/"><img src="${base64Img}"></a></p>`
-        );
+        await tick();
+        expect(".o_we_apply_link").toHaveAttribute("disabled");
     });
 });
 
-describe("label is a valid URL", () => {
+describe("popover with a valid URL as label", () => {
     test("URL input should be prefilled by corresponding URL if the label matches the URL format", async () => {
         const { el } = await setupEditor("<p>[google.com]</p>");
         await waitFor(".o-we-toolbar");
@@ -1966,21 +1031,4 @@ describe("label is a valid URL", () => {
             message: "should focus label input by default, when we don't have a label",
         });
     });
-});
-
-test("Should properly show the preview if fetching metadata fails", async () => {
-    const id = Math.random().toString();
-    onRpc("/html_editor/link_preview_internal", () => Promise.reject(new Error(`No data ${id}`)));
-    onRpc("/contactus", () => ({}));
-    const originalConsoleWarn = console.warn.bind(console);
-    patchWithCleanup(console, {
-        warn: (msg, error, ...args) => {
-            if (!error?.message?.includes?.(id)) {
-                originalConsoleWarn(msg, error, ...args);
-            }
-        },
-    });
-    const { el } = await setupEditor('<p><a href="/contactus">a[]b</a></p>');
-    await waitFor(".o-we-linkpopover");
-    expect(cleanLinkArtifacts(getContent(el))).toBe('<p><a href="/contactus">a[]b</a></p>');
 });
