+++ addons/html_editor/static/tests/link/transform.test.js
@@ -1,4 +1,4 @@
-import { expect, test } from "@odoo/hoot";
+import { describe, expect, test } from "@odoo/hoot";
 import { manuallyDispatchProgrammaticEvent } from "@odoo/hoot-dom";
 import { tick } from "@odoo/hoot-mock";
 import { onRpc, patchWithCleanup } from "@web/../tests/web_test_helpers";
@@ -8,164 +8,199 @@ import { getContent, setSelection } from "../_helpers/selection";
 import { ensureDistinctHistoryStep, insertSpace, insertText, undo } from "../_helpers/user_actions";
 import { expectElementCount } from "../_helpers/ui_expectations";
 
-/**
- * Automatic link creation when pressing Space, Enter or Shift+Enter after an url
- */
-test("should transform url after space (1)", async () => {
-    await testEditor({
-        contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
-        stepFunction: async (editor) => {
-            await insertSpace(editor);
-        },
-        contentAfter:
-            '<p>a http://test.com b <a href="http://test.com">http://test.com</a>&nbsp;[] c http://test.com d</p>',
+describe("transform on space", () => {
+    /**
+     * Automatic link creation when pressing Space, Enter or Shift+Enter after an url
+     */
+    test("should transform url after space (1)", async () => {
+        await testEditor({
+            contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
+            stepFunction: async (editor) => {
+                await insertSpace(editor);
+            },
+            contentAfter:
+                '<p>a http://test.com b <a href="http://test.com">http://test.com</a>&nbsp;[] c http://test.com d</p>',
+        });
+    });
+    test("should transform url after space (2)", async () => {
+        await testEditor({
+            contentBefore: "<p>http://test.com[]</p>",
+            stepFunction: async (editor) => {
+                // Setup: simulate multiple text nodes in a p: <p>"http://test" ".com"</p>
+                editor.editable.firstChild.firstChild.splitText(11);
+
+                /** @todo fix warnings */
+                patchWithCleanup(console, { warn: () => {} });
+
+                // Action: insert space
+                await insertSpace(editor);
+            },
+            contentAfter: '<p><a href="http://test.com">http://test.com</a>&nbsp;[]</p>',
+        });
+    });
+
+    test("typing valid URL + space should convert to link", async () => {
+        const { editor, el } = await setupEditor("<p>[]</p>");
+        await insertText(editor, "http://google.co.in");
+        await insertText(editor, " ");
+        expect(cleanLinkArtifacts(getContent(el))).toBe(
+            '<p><a href="http://google.co.in">http://google.co.in</a>&nbsp;[]</p>'
+        );
+    });
+    test("typing valid URL without protocol + space should convert to https link", async () => {
+        const { editor, el } = await setupEditor("<p>[]</p>");
+        await insertText(editor, "google.com");
+        await insertText(editor, " ");
+        expect(cleanLinkArtifacts(getContent(el))).toBe(
+            '<p><a href="https://google.com">google.com</a>&nbsp;[]</p>'
+        );
+    });
+    test("typing valid http URL + space should convert to http link", async () => {
+        const { editor, el } = await setupEditor("<p>[]</p>");
+        await insertText(editor, "http://google.com");
+        await insertText(editor, " ");
+        expect(cleanLinkArtifacts(getContent(el))).toBe(
+            '<p><a href="http://google.com">http://google.com</a>&nbsp;[]</p>'
+        );
+    });
+    test("typing invalid URL + space should not convert to link", async () => {
+        const { editor, el } = await setupEditor("<p>[]</p>");
+        await insertText(editor, "www.odoo");
+        await insertText(editor, " ");
+        expect(cleanLinkArtifacts(getContent(el))).toBe("<p>www.odoo []</p>");
+    });
+
+    test("should transform url followed by punctuation characters after space (1)", async () => {
+        await testEditor({
+            contentBefore: "<p>http://test.com.[]</p>",
+            stepFunction: async (editor) => {
+                await insertSpace(editor);
+            },
+            contentAfter: '<p><a href="http://test.com">http://test.com</a>.&nbsp;[]</p>',
+        });
+    });
+    test("should transform url followed by punctuation characters after space (2)", async () => {
+        await testEditor({
+            contentBefore: "<p>test.com...[]</p>",
+            stepFunction: (editor) => insertSpace(editor),
+            contentAfter: '<p><a href="https://test.com">test.com</a>...&nbsp;[]</p>',
+        });
+    });
+    test("should transform url followed by punctuation characters after space (3)", async () => {
+        await testEditor({
+            contentBefore: "<p>test.com,[]</p>",
+            stepFunction: (editor) => insertSpace(editor),
+            contentAfter: '<p><a href="https://test.com">test.com</a>,&nbsp;[]</p>',
+        });
+    });
+    test("should transform url followed by punctuation characters after space (4)", async () => {
+        await testEditor({
+            contentBefore: "<p>test.com,hello[]</p>",
+            stepFunction: (editor) => insertSpace(editor),
+            contentAfter: '<p><a href="https://test.com">test.com</a>,hello&nbsp;[]</p>',
+        });
+    });
+    test("should transform url followed by punctuation characters after space (5)", async () => {
+        await testEditor({
+            contentBefore: "<p>http://test.com[]</p>",
+            stepFunction: async (editor) => {
+                // Setup: simulate multiple text nodes in a p: <p>"http://test" ".com"</p>
+                editor.editable.firstChild.firstChild.splitText(11);
+
+                /** @todo fix warnings */
+                patchWithCleanup(console, { warn: () => {} });
+
+                // Action: insert space
+                await insertSpace(editor);
+            },
+            contentAfter: '<p><a href="http://test.com">http://test.com</a>&nbsp;[]</p>',
+        });
+    });
+    test("transform text url into link and undo it", async () => {
+        const { el, editor } = await setupEditor(`<p>[]</p>`);
+        await insertText(editor, "www.abc.jpg");
+        await ensureDistinctHistoryStep();
+        await insertText(editor, " ");
+        expect(cleanLinkArtifacts(getContent(el))).toBe(
+            '<p><a href="https://www.abc.jpg">www.abc.jpg</a>&nbsp;[]</p>'
+        );
+
+        undo(editor);
+        expect(cleanLinkArtifacts(getContent(el))).toBe(
+            '<p><a href="https://www.abc.jpg">www.abc.jpg</a>[]</p>'
+        );
+
+        undo(editor);
+        expect(cleanLinkArtifacts(getContent(el))).toBe("<p>www.abc.jpg[]</p>");
+    });
+
+    test("should show replace URL button if link is created by transformation", async () => {
+        onRpc("/html_editor/link_preview_external", () => ({
+            og_description:
+                "From ERP to CRM, eCommerce and CMS. Download Odoo or use it in the cloud. Grow Your Business.",
+            og_image: "https://www.odoo.com/web/image/41207129-1abe7a15/homepage-seo.png",
+            og_title: "Open Source ERP and CRM | Odoo",
+            og_type: "website",
+            og_site_name: "Odoo",
+            source_url: "https://odoo.com",
+        }));
+        const { editor } = await setupEditor(`<p>[]</p>`);
+        await insertText(editor, "https://odoo.com ");
+        const link = document.querySelector("a");
+        setSelection({
+            anchorNode: link,
+            anchorOffset: 0,
+        });
+        await tick();
+        await expectElementCount(".o-we-linkpopover", 1);
+        expect("button.o_we_replace_title_btn").toHaveCount(1);
     });
 });
-test("should transform url after space (2)", async () => {
-    await testEditor({
-        contentBefore: "<p>http://test.com[]</p>",
-        stepFunction: async (editor) => {
-            // Setup: simulate multiple text nodes in a p: <p>"http://test" ".com"</p>
-            editor.editable.firstChild.firstChild.splitText(11);
 
-            /** @todo fix warnings */
-            patchWithCleanup(console, { warn: () => {} });
+describe("transform on enter and shift+enter", () => {
+    test("should transform url after enter", async () => {
+        await testEditor({
+            contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
+            stepFunction: async (editor) => {
+                // Simulate "Enter"
+                await manuallyDispatchProgrammaticEvent(editor.editable, "beforeinput", {
+                    inputType: "insertParagraph",
+                });
+            },
+            contentAfter:
+                '<p>a http://test.com b <a href="http://test.com">http://test.com</a></p><p>[]&nbsp;c http://test.com d</p>',
+        });
+    });
 
-            // Action: insert space
-            await insertSpace(editor);
-        },
-        contentAfter: '<p><a href="http://test.com">http://test.com</a>&nbsp;[]</p>',
+    test("should transform url after shift+enter", async () => {
+        await testEditor({
+            contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
+            stepFunction: async (editor) => {
+                // Simulate "Shift + Enter"
+                await manuallyDispatchProgrammaticEvent(editor.editable, "beforeinput", {
+                    inputType: "insertLineBreak",
+                });
+            },
+            contentAfter:
+                '<p>a http://test.com b <a href="http://test.com">http://test.com</a><br>[]&nbsp;c http://test.com d</p>',
+        });
     });
 });
 
-test("should transform url followed by punctuation characters after space (1)", async () => {
-    await testEditor({
-        contentBefore: "<p>http://test.com.[]</p>",
-        stepFunction: async (editor) => {
-            await insertSpace(editor);
-        },
-        contentAfter: '<p><a href="http://test.com">http://test.com</a>.&nbsp;[]</p>',
+describe("should not transform url", () => {
+    test("should not transform an email url after space", async () => {
+        await testEditor({
+            contentBefore: "<p>user@domain.com[]</p>",
+            stepFunction: (editor) => insertSpace(editor),
+            contentAfter: "<p>user@domain.com&nbsp;[]</p>",
+        });
     });
-});
-test("should transform url followed by punctuation characters after space (2)", async () => {
-    await testEditor({
-        contentBefore: "<p>test.com...[]</p>",
-        stepFunction: (editor) => insertSpace(editor),
-        contentAfter: '<p><a href="https://test.com">test.com</a>...&nbsp;[]</p>',
+    test("should not transform url after two space", async () => {
+        await testEditor({
+            contentBefore: "<p>a http://test.com b http://test.com&nbsp;[] c http://test.com d</p>",
+            stepFunction: (editor) => insertSpace(editor),
+            contentAfter:
+                "<p>a http://test.com b http://test.com&nbsp; []&nbsp;c http://test.com d</p>",
+        });
     });
-});
-test("should transform url followed by punctuation characters after space (3)", async () => {
-    await testEditor({
-        contentBefore: "<p>test.com,[]</p>",
-        stepFunction: (editor) => insertSpace(editor),
-        contentAfter: '<p><a href="https://test.com">test.com</a>,&nbsp;[]</p>',
-    });
-});
-test("should transform url followed by punctuation characters after space (4)", async () => {
-    await testEditor({
-        contentBefore: "<p>test.com,hello[]</p>",
-        stepFunction: (editor) => insertSpace(editor),
-        contentAfter: '<p><a href="https://test.com">test.com</a>,hello&nbsp;[]</p>',
-    });
-});
-test("should transform url followed by punctuation characters after space (5)", async () => {
-    await testEditor({
-        contentBefore: "<p>http://test.com[]</p>",
-        stepFunction: async (editor) => {
-            // Setup: simulate multiple text nodes in a p: <p>"http://test" ".com"</p>
-            editor.editable.firstChild.firstChild.splitText(11);
-
-            /** @todo fix warnings */
-            patchWithCleanup(console, { warn: () => {} });
-
-            // Action: insert space
-            await insertSpace(editor);
-        },
-        contentAfter: '<p><a href="http://test.com">http://test.com</a>&nbsp;[]</p>',
-    });
-});
-
-test("should transform url after enter", async () => {
-    await testEditor({
-        contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
-        stepFunction: async (editor) => {
-            // Simulate "Enter"
-            await manuallyDispatchProgrammaticEvent(editor.editable, "beforeinput", {
-                inputType: "insertParagraph",
-            });
-        },
-        contentAfter:
-            '<p>a http://test.com b <a href="http://test.com">http://test.com</a></p><p>[]&nbsp;c http://test.com d</p>',
-    });
-});
-
-test("should transform url after shift+enter", async () => {
-    await testEditor({
-        contentBefore: "<p>a http://test.com b http://test.com[] c http://test.com d</p>",
-        stepFunction: async (editor) => {
-            // Simulate "Shift + Enter"
-            await manuallyDispatchProgrammaticEvent(editor.editable, "beforeinput", {
-                inputType: "insertLineBreak",
-            });
-        },
-        contentAfter:
-            '<p>a http://test.com b <a href="http://test.com">http://test.com</a><br>[]&nbsp;c http://test.com d</p>',
-    });
-});
-
-test("should not transform an email url after space", async () => {
-    await testEditor({
-        contentBefore: "<p>user@domain.com[]</p>",
-        stepFunction: (editor) => insertSpace(editor),
-        contentAfter: "<p>user@domain.com&nbsp;[]</p>",
-    });
-});
-
-test("should not transform url after two space", async () => {
-    await testEditor({
-        contentBefore: "<p>a http://test.com b http://test.com&nbsp;[] c http://test.com d</p>",
-        stepFunction: (editor) => insertSpace(editor),
-        contentAfter:
-            "<p>a http://test.com b http://test.com&nbsp; []&nbsp;c http://test.com d</p>",
-    });
-});
-
-test("transform text url into link and undo it", async () => {
-    const { el, editor } = await setupEditor(`<p>[]</p>`);
-    await insertText(editor, "www.abc.jpg");
-    await ensureDistinctHistoryStep();
-    await insertText(editor, " ");
-    expect(cleanLinkArtifacts(getContent(el))).toBe(
-        '<p><a href="https://www.abc.jpg">www.abc.jpg</a>&nbsp;[]</p>'
-    );
-
-    undo(editor);
-    expect(cleanLinkArtifacts(getContent(el))).toBe(
-        '<p><a href="https://www.abc.jpg">www.abc.jpg</a>[]</p>'
-    );
-
-    undo(editor);
-    expect(cleanLinkArtifacts(getContent(el))).toBe("<p>www.abc.jpg[]</p>");
-});
-
-test("should show replace URL button if link is created by transformation", async () => {
-    onRpc("/html_editor/link_preview_external", () => ({
-        og_description:
-            "From ERP to CRM, eCommerce and CMS. Download Odoo or use it in the cloud. Grow Your Business.",
-        og_image: "https://www.odoo.com/web/image/41207129-1abe7a15/homepage-seo.png",
-        og_title: "Open Source ERP and CRM | Odoo",
-        og_type: "website",
-        og_site_name: "Odoo",
-        source_url: "https://odoo.com",
-    }));
-    const { editor } = await setupEditor(`<p>[]</p>`);
-    await insertText(editor, "https://odoo.com ");
-    const link = document.querySelector("a");
-    setSelection({
-        anchorNode: link,
-        anchorOffset: 0,
-    });
-    await tick();
-    await expectElementCount(".o-we-linkpopover", 1);
-    expect("button.o_we_replace_title_btn").toHaveCount(1);
 });
